idk why my squares are not moving:(

+++ src/objects/Square.h
@@ -20,8 +20,8 @@ public:
   ~Square();
 
   void draw() override;
-  void moveHorizontally(int delta_x) override;
-  void moveVertically(int delta_y) override;
+  void moveHorizontally(float delta_x) override;
+  void moveVertically(float delta_y) override;
   void rotateClockwise();
   void rotateCounterClockwise();
